Added protection to the routes based on user token
