--- UIApp/src/components/SavedSearch/SavedSearch.js
@@ -1,94 +1,0 @@
-import React from "react";
-import { useState } from 'react';
-import MaterialTable from 'material-table';
-import { data } from '../FavSearchData';
-import { forwardRef } from "react";
-import AddBox from '@material-ui/icons/AddBox';
-import ArrowDownward from '@material-ui/icons/ArrowDownward';
-import Check from '@material-ui/icons/Check';
-import ChevronRight from '@material-ui/icons/ChevronRight';
-import Clear from '@material-ui/icons/Clear';
-import DeleteOutline from '@material-ui/icons/DeleteOutline';
-import EditIcon from '@material-ui/icons/Edit';
-import FilterList from '@material-ui/icons/FilterList';
-import Remove from '@material-ui/icons/Remove';
-import SaveAlt from '@material-ui/icons/SaveAlt';
-import Search from '@material-ui/icons/Search';
-import ViewColumn from '@material-ui/icons/ViewColumn';
-
-const tableIcons = {
-    Add: forwardRef((props, ref) => <AddBox {...props} ref={ref} />),
-    Check: forwardRef((props, ref) => <Check {...props} ref={ref} />),
-    Clear: forwardRef((props, ref) => <Clear {...props} ref={ref} />),
-    Delete: forwardRef((props, ref) => <DeleteOutline {...props} ref={ref} />),
-    DetailPanel: forwardRef((props, ref) => <ChevronRight {...props} ref={ref} />),
-    Edit: forwardRef((props, ref) => <EditIcon {...props} ref={ref} />),
-    Export: forwardRef((props, ref) => <SaveAlt {...props} ref={ref} />),
-    Filter: forwardRef((props, ref) => <FilterList {...props} ref={ref} />),
-    ResetSearch: forwardRef((props, ref) => <Clear {...props} ref={ref} />),
-    Search: forwardRef((props, ref) => <Search {...props} ref={ref} />),
-    SortArrow: forwardRef((props, ref) => <ArrowDownward {...props} ref={ref} />),
-    ThirdStateCheck: forwardRef((props, ref) => <Remove {...props} ref={ref} />),
-    ViewColumn: forwardRef((props, ref) => <ViewColumn {...props} ref={ref} />)
-  };
-
-
-export default function Editable() {
-  const { useState } = React;
-  const columns = [
-    { title: 'Search name', field: 'searchName' },
-    { title: 'Location', field: 'location' }
-  ];
-
-  const [dataMain, setData] = useState([...data
-  ]);
-
-  return (
-    <div style={{ maxWidth: '95%' }}>
-    <MaterialTable
-    title='Favorite search by job title'
-       localization={{
-        header: {
-            actions: 'Edit Delete'
-        }
-    }}
-      columns={columns}
-      icons={tableIcons}
-      options={{
-        actionsColumnIndex: -1,
-
-        showFirstLastPageButtons: false,
-        paging: false
-
-      }}
-      data={dataMain}
-      editable={{
-        onRowUpdate: (newData, oldData) =>
-          new Promise((resolve, reject) => {
-            setTimeout(() => {
-              const dataUpdate = [...data];
-              const index = oldData.tableData.id;
-              dataUpdate[index] = newData;
-              setData([...dataUpdate]);
-              resolve();
-            }, 1000)
-          }),
-        onRowDelete: oldData =>
-          new Promise((resolve, reject) => {
-            setTimeout(() => {
-              const dataDelete = [...data];
-              const index = oldData.tableData.id;
-              dataDelete.splice(index, 1);
-              setData([...dataDelete]);
-              
-              resolve()
-            }, 1000)
-          }),
-      }}
-    />
-    </div>
-  )
-}
-
-
-
